requirements.txt
aniso8601==9.0.1
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
cx_Oracle==8.3.0
distlib==0.3.8
filelock==3.13.1
Flask==3.0.0
Flask-RESTful==0.3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oracledb==2.0.0
pipenv==2023.11.15
platformdirs==4.1.0
pycparser==2.21
python-dotenv==1.0.0
pytz==2023.3.post1
setuptools==69.0.3
six==1.16.0
virtualenv==20.25.0
Werkzeug==3.0.1
wheel==0.42.0
Flask-HTTPAuth==4.8.0